32200 with 32541

+++ code/__defines/_renderer.dm
@@ -78,20 +78,29 @@ INITIALIZE_IMMEDIATE(/atom/movable/renderer)
 	if (!renderers)
 		renderers = list()
 	for (var/atom/movable/renderer/renderer as anything in subtypesof(/atom/movable/renderer))
+		if(ispath(renderer, /atom/movable/renderer/shared))
+			continue
 		renderer = new renderer (null, src)
 		renderers[renderer] = renderer.plane // (renderer = plane) format for visual debugging
 		if (renderer.relay)
 			my_client.screen += renderer.relay
 		my_client.screen += renderer
+	for (var/atom/movable/renderer/zrenderer as anything in GLOB.zmimic_renderers)
+		if (zrenderer.relay)
+			my_client.screen += zrenderer.relay
+		my_client.screen += zrenderer
 
 
 /// Removes the mob's renderers on /Logout()
 /mob/proc/RemoveRenderers()
-	for(var/atom/movable/renderer/renderer as anything in renderers)
-		my_client.screen -= renderer
-		if (renderer.relay)
-			my_client.screen -= renderer.relay
-		qdel(renderer)
+	if(my_client)
+		for(var/atom/movable/renderer/renderer as anything in renderers)
+			my_client.screen -= renderer
+			if (renderer.relay)
+				my_client.screen -= renderer.relay
+			qdel(renderer)
+		for (var/atom/movable/renderer/renderer as anything in GLOB.zmimic_renderers)
+			my_client.screen -= renderer
 	if (renderers)
 		renderers.Cut()
 
@@ -114,6 +123,39 @@ INITIALIZE_IMMEDIATE(/atom/movable/renderer)
 	appearance_flags = PLANE_MASTER | NO_CLIENT_COLOR
 	blend_mode = BLEND_MULTIPLY
 	color = list(null, null, null, "#0000", "#000f")
+
+
+/atom/movable/renderer/space
+	name = "Space"
+	group = RENDER_GROUP_SCENE
+	plane = SPACE_PLANE
+
+
+/atom/movable/renderer/skybox
+	name = "Skybox"
+	group = RENDER_GROUP_SCENE
+	plane = SKYBOX_PLANE
+	relay_blend_mode = BLEND_MULTIPLY
+
+
+//Z Mimic planemasters -> Could apply scaling for parallax though that requires copying appearances from adjacent turfs
+GLOBAL_LIST_EMPTY(zmimic_renderers)
+
+/hook/startup/proc/create_global_renderers() //Some (most) renderers probably do not need to be instantiated per mob. So may as well make them global and just add to screen
+	//Zmimic planemasters
+	for(var/i = 0 to OPENTURF_MAX_DEPTH)
+		GLOB.zmimic_renderers += new /atom/movable/renderer/shared/zmimic(null, null, OPENTURF_MAX_PLANE - i)
+
+	return TRUE
+
+/atom/movable/renderer/shared/zmimic
+	name = "Zrenderer"
+	group = RENDER_GROUP_SCENE
+
+/atom/movable/renderer/shared/zmimic/Initialize(mapload, _owner, _plane)
+	plane = _plane
+	name = "Zrenderer [plane]"
+	. = ..()
 
 
 // Draws the game world; live mobs, items, turfs, etc.
